packages/modules/AdServices/adservices/tests/unittest/service-core/src/com/android/adservices/data/common/CleanupUtilsTest.java
/*
 * Copyright (C) 2023 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.android.adservices.data.common;

import static com.android.dx.mockito.inline.extended.ExtendedMockito.any;
import static com.android.dx.mockito.inline.extended.ExtendedMockito.anyInt;
import static com.android.dx.mockito.inline.extended.ExtendedMockito.doReturn;

import static org.junit.Assert.assertEquals;

import android.adservices.common.CommonFixture;
import android.content.Context;
import android.content.pm.ApplicationInfo;

import androidx.test.core.app.ApplicationProvider;

import com.android.adservices.service.Flags;
import com.android.adservices.service.FlagsFactory;
import com.android.adservices.service.common.AllowLists;
import com.android.adservices.service.common.compat.PackageManagerCompatUtils;
import com.android.dx.mockito.inline.extended.ExtendedMockito;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.mockito.MockitoSession;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

public class CleanupUtilsTest {
    private static final Context CONTEXT = ApplicationProvider.getApplicationContext();
    private MockitoSession mStaticMockSession = null;

    @Before
    public void setup() {
        mStaticMockSession =
                ExtendedMockito.mockitoSession()
                        .spyStatic(FlagsFactory.class)
                        .mockStatic(PackageManagerCompatUtils.class)
                        .initMocks(this)
                        .startMocking();
    }

    @After
    public void teardown() {
        if (mStaticMockSession != null) {
            mStaticMockSession.finishMocking();
        }
    }

    @Test
    public void testEmpty() {
        class FlagsThatAllowOneApp implements Flags {
            @Override
            public String getPpapiAppAllowList() {
                return CommonFixture.TEST_PACKAGE_NAME_1;
            }
        }
        List<String> packageList = new ArrayList<>();
        CleanupUtils.removeAllowedPackages(
                packageList, CONTEXT.getPackageManager(), new FlagsThatAllowOneApp());
        assertEquals(new ArrayList<>(), packageList);
    }

    @Test
    public void testCleanupNotUninstalled() {
        class FlagsThatAllowOneApp implements Flags {
            @Override
            public String getPpapiAppAllowList() {
                return CommonFixture.TEST_PACKAGE_NAME_1;
            }
        }

        ApplicationInfo installedPackage1 = new ApplicationInfo();
        installedPackage1.packageName = CommonFixture.TEST_PACKAGE_NAME_1;
        ApplicationInfo installedPackage2 = new ApplicationInfo();
        installedPackage2.packageName = CommonFixture.TEST_PACKAGE_NAME_2;
        doReturn(Arrays.asList(installedPackage1, installedPackage2))
                .when(() -> PackageManagerCompatUtils.getInstalledApplications(any(), anyInt()));
        List<String> packageList =
                new ArrayList<>(
                        Arrays.asList(
                                CommonFixture.TEST_PACKAGE_NAME_1,
                                CommonFixture.TEST_PACKAGE_NAME_2));
        List<String> expected = Arrays.asList(CommonFixture.TEST_PACKAGE_NAME_2);
        CleanupUtils.removeAllowedPackages(
                packageList, CONTEXT.getPackageManager(), new FlagsThatAllowOneApp());
        assertEquals(expected, packageList);
    }

    @Test
    public void testCleanupNotAllowed() {
        // All owners are allowed
        class FlagsThatAllowAllApps implements Flags {
            @Override
            public String getPpapiAppAllowList() {
                return AllowLists.ALLOW_ALL;
            }
        }

        ApplicationInfo installedPackage2 = new ApplicationInfo();
        installedPackage2.packageName = CommonFixture.TEST_PACKAGE_NAME_2;
        doReturn(Arrays.asList(installedPackage2))
                .when(() -> PackageManagerCompatUtils.getInstalledApplications(any(), anyInt()));
        List<String> packageList =
                new ArrayList<>(
                        Arrays.asList(
                                CommonFixture.TEST_PACKAGE_NAME_1,
                                CommonFixture.TEST_PACKAGE_NAME_2));
        List<String> expected = Arrays.asList(CommonFixture.TEST_PACKAGE_NAME_1);
        CleanupUtils.removeAllowedPackages(
                packageList, CONTEXT.getPackageManager(), new FlagsThatAllowAllApps());
        assertEquals(expected, packageList);
    }
}
